Move hooks and lib folder outside the Payment method overview folder.
