+++ client/a8c-for-agencies/sections/purchases/payment-methods/lib/get-stripe-configuration.ts
@@ -4,6 +4,7 @@ import type { GetStripeConfigurationArgs } from '@automattic/calypso-stripe';
 export async function getStripeConfiguration(
 	requestArgs: GetStripeConfigurationArgs & { needs_intent?: boolean }
 ) {
+	// FIXME: Update this to the correct endpoint.
 	return await wp.req.get(
 		{
 			apiNamespace: 'wpcom/v2',
