+++ client/a8c-for-agencies/sections/purchases/payment-methods/payment-method-overview/stored-credit-card/index.tsx
@@ -3,8 +3,8 @@ import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import { useDeleteCard } from '../hooks/use-delete-card';
-import { useSetAsPrimaryCard } from '../hooks/use-set-as-primary-card';
+import { useDeleteCard } from '../../hooks/use-delete-card';
+import { useSetAsPrimaryCard } from '../../hooks/use-set-as-primary-card';
 import StoredCreditCardDeleteDialog from '../stored-credit-card-delete-dialog';
 import CreditCardActions from './credit-card-actions';
 import type { PaymentMethod } from 'calypso/jetpack-cloud/sections/partner-portal/payment-methods';
